+++ Models/ComprasModel.cs
@@ -13,10 +13,11 @@ namespace TesteUGB.Models
         public DateTime DataSolicitada { get; set; }
         public DateTime DataPrevisaoEntregaProduto { get; set; }
         public TipoDoProdutoEnum TipoDoProduto { get; set; }
-        public int ValorUnitarioDoProduto { get; set; }
-        public int ValorTotal { get; set; } // Soma do valor unitario x quantidade
+        public double ValorUnitarioDoProduto { get; set; }
+        public double ValorTotal { get; set; } // Soma do valor unitario x quantidade
         public long NumeroNotaFiscalProduto { get; set; }
         public long CodigoEAN { get; set; } // Código de Barras do Produto
+        //public byte[] Conteudo { get; set; } // gerar pdf
     }
 }
